services/web-app/requirements.txt
# Web framework
Flask==2.3.3

# Kafka/RedPanda client
kafka-python==2.0.2

# Configuration and environment
python-dotenv==1.0.0
